Add basic Spotify functionality
This commit adds a new app, Spotify.r, which can be used to make a map from a playlist URI. Changes to utils.r include helper functions for achieving this, and a notebook to show the process of creating these functions.

Next commit should merge spotify.r and app.r, or fully separate them into different Shiny pages, but at the moment there is a lot of overlap.